Fix Firebase Auth persistence and reorganize account module
- Fix Firebase Auth React Native persistence by using official
  getReactNativePersistence from firebase/auth instead of custom implementation
- Fix navigation errors (GO_BACK/POP not handled) by using router.navigate('/(tabs)')
  instead of router.back()/router.dismiss()
- Reorganize account feature into app/(account)/ group folder with dedicated
  _layout.tsx and _components/ subfolder for better code organization
- Update all account component imports to new locations
- Simplify account screen navigation flow
