+++ app/(account)/_layout.tsx
@@ -1,0 +1,10 @@
+import { Stack } from 'expo-router';
+
+export default function AccountLayout() {
+  return (
+    <Stack screenOptions={{ headerShown: false }}>
+      <Stack.Screen name="edit-profile" />
+      <Stack.Screen name="change-password" />
+    </Stack>
+  );
+}
