+++ app/(account)/edit-profile.tsx
@@ -1,7 +1,6 @@
 import {
   StyleSheet,
   ScrollView,
-  View,
   Pressable,
   Alert,
 } from 'react-native';
@@ -10,36 +9,28 @@ import { useRouter } from 'expo-router';
 import { ThemedText } from '@/components/themed-text';
 import { ThemedView } from '@/components/themed-view';
 import { useColorScheme } from '@/hooks/use-color-scheme';
-import { ChangePasswordForm } from '@spezivibe/account';
+import { EditProfileForm } from '@spezivibe/account';
 import { Ionicons } from '@expo/vector-icons';
 
-export default function ChangePasswordScreen() {
+export default function EditProfileScreen() {
   const router = useRouter();
   const colorScheme = useColorScheme();
   const isDark = colorScheme === 'dark';
 
-  const handleSuccess = () => {
-    Alert.alert('Success', 'Password changed successfully', [
-      {
-        text: 'OK',
-        onPress: () => router.back(),
-      },
-    ]);
-  };
+  const handleClose = () => router.navigate('/(tabs)');
 
   const handleError = (error: Error) => {
-    console.error('Error changing password:', error);
-    Alert.alert('Error', error.message || 'Failed to change password. Please try again.');
+    Alert.alert('Error', error.message || 'Failed to update profile. Please try again.');
   };
 
   return (
     <SafeAreaView style={[styles.safeArea, { backgroundColor: isDark ? '#000' : '#f2f2f7' }]}>
       <ThemedView style={styles.header}>
         <ThemedText type="title" style={styles.headerTitle}>
-          Change Password
+          Edit Profile
         </ThemedText>
         <Pressable
-          onPress={() => router.back()}
+          onPress={handleClose}
           style={({ pressed }) => [
             styles.closeButton,
             { opacity: pressed ? 0.5 : 1 },
@@ -61,29 +52,8 @@ export default function ChangePasswordScreen() {
         keyboardShouldPersistTaps="handled"
       >
         <ThemedView style={styles.container}>
-          {/* Information Card */}
-          <View
-            style={[
-              styles.infoCard,
-              {
-                backgroundColor: isDark ? 'rgba(184, 58, 75, 0.15)' : 'rgba(140, 21, 21, 0.1)',
-                borderColor: isDark ? 'rgba(184, 58, 75, 0.3)' : 'rgba(140, 21, 21, 0.2)',
-              },
-            ]}
-          >
-            <Ionicons
-              name="information-circle"
-              size={24}
-              color={isDark ? '#B83A4B' : '#8C1515'}
-              style={styles.infoIcon}
-            />
-            <ThemedText style={styles.infoText}>
-              For your security, please enter your current password to make changes.
-            </ThemedText>
-          </View>
-
-          <ChangePasswordForm
-            onSuccess={handleSuccess}
+          <EditProfileForm
+            onSuccess={handleClose}
             onError={handleError}
             containerStyle={styles.form}
             inputStyle={[
@@ -103,8 +73,7 @@ export default function ChangePasswordScreen() {
               styles.saveButtonText,
               { color: isDark ? '#000' : '#fff' },
             ]}
-            buttonText="Change Password"
-            showRequirements={true}
+            buttonText="Save Changes"
           />
         </ThemedView>
       </ScrollView>
@@ -141,23 +110,6 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     padding: 20,
-  },
-  infoCard: {
-    flexDirection: 'row',
-    padding: 16,
-    borderRadius: 12,
-    marginBottom: 20,
-    borderWidth: 1,
-  },
-  infoIcon: {
-    marginRight: 12,
-    marginTop: 2,
-  },
-  infoText: {
-    flex: 1,
-    fontSize: 14,
-    lineHeight: 20,
-    opacity: 0.9,
   },
   form: {
     width: '100%',
